cleanup BoxSelector trait(move methods out);
