+++ ergo-wallet/src/main/scala/org/ergoplatform/wallet/boxes/BoxSelector.scala
@@ -1,7 +1,6 @@
 package org.ergoplatform.wallet.boxes
 
 import org.ergoplatform.ErgoBoxAssets
-import org.ergoplatform.ErgoBoxAssetsHolder
 import org.ergoplatform.ErgoBox.MaxTokens
 import org.ergoplatform.wallet.boxes.BoxSelector.BoxSelectionResult
 import scorex.util.ModifierId
@@ -38,7 +37,14 @@ trait BoxSelector {
   ): Option[BoxSelectionResult[T]] =
     select(inputBoxes, _ => true, targetBalance, targetAssets)
 
-  protected  def mergeAssetsMut(
+
+}
+
+object BoxSelector {
+
+  final case class BoxSelectionResult[T <: ErgoBoxAssets](boxes: Seq[T], changeBoxes: Seq[ErgoBoxAssets])
+
+  def mergeAssetsMut(
     into: mutable.Map[ModifierId, Long],
     from: Map[ModifierId, Long]*
   ): Unit = {
@@ -48,7 +54,7 @@ trait BoxSelector {
     })
   }
 
-  protected  def subtractAssetsMut(
+  def subtractAssetsMut(
     from: mutable.Map[ModifierId, Long],
     subtractor: Map[ModifierId, Long]
   ): Unit = {
@@ -62,59 +68,5 @@ trait BoxSelector {
         }
     }
   }
-
-  protected def calcChange[T <: ErgoBoxAssets](
-    boxes: Seq[T],
-    targetBalance: Long,
-    targetAssets: Map[ModifierId, Long]
-  ): Option[Seq[ErgoBoxAssets]] = {
-    val compactedBalance = boxes.map(_.value).sum
-    val compactedAssets  = mutable.Map[ModifierId, Long]()
-    mergeAssetsMut(compactedAssets, boxes.map(_.tokens): _*)
-
-    subtractAssetsMut(compactedAssets, targetAssets)
-    val changeBoxesAssets: Seq[mutable.Map[ModifierId, Long]] =
-      compactedAssets.grouped(MaxTokens).toSeq
-    val changeBalance = compactedBalance - targetBalance
-    formChangeBoxes(changeBalance, changeBoxesAssets)
-  }
-
-  protected def formChangeBoxes(
-    changeBalance: Long,
-    changeBoxesAssets: Seq[mutable.Map[ModifierId, Long]]
-  ): Option[Seq[ErgoBoxAssets]] = {
-    //at least 1 ergo token should be assigned per a created box
-    if (changeBoxesAssets.size > changeBalance) {
-      None
-    } else {
-      val changeBoxes = if (changeBoxesAssets.nonEmpty) {
-        val baseChangeBalance = changeBalance / changeBoxesAssets.size
-
-        val changeBoxesNoBalanceAdjusted = changeBoxesAssets.map { a =>
-          ErgoBoxAssetsHolder(baseChangeBalance, a.toMap)
-        }
-
-        val modifiedBoxOpt = changeBoxesNoBalanceAdjusted.headOption.map { firstBox =>
-          ErgoBoxAssetsHolder(
-            changeBalance - baseChangeBalance * (changeBoxesAssets.size - 1),
-            firstBox.tokens
-          )
-        }
-
-        modifiedBoxOpt.toSeq ++ changeBoxesNoBalanceAdjusted.tail
-      } else if (changeBalance > 0) {
-        Seq(ErgoBoxAssetsHolder(changeBalance))
-      } else {
-        Seq.empty
-      }
-      Some(changeBoxes)
-    }
-  }
-
-}
-
-object BoxSelector {
-
-  final case class BoxSelectionResult[T <: ErgoBoxAssets](boxes: Seq[T], changeBoxes: Seq[ErgoBoxAssets])
 
 }
